Test: active resources working

+++ src/main/java/com/automation/api/models/Resource.java
@@ -14,7 +14,7 @@ public class Resource {
     private String name;
     private String trademark;
     private int stock;
-    private double price;
+    private float price;
     private String description;
     private String tags;
     private boolean active;
